requirements.txt
# Core data processing and analysis
polars>=0.20.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.17.0
jupyter>=1.0.0
notebook>=7.0.0

# API access for data collection
fredapi>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0

# Time series analysis and forecasting
statsmodels>=0.14.0
scipy>=1.11.0
scikit-learn>=1.3.0
xgboost>=2.0.0
holidays>=0.34

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.66.0